StarNames 파일 만들어서 dictionary 추가 / UILabel 함수 인자 추가
