+++ polaris-ios/polaris-ios/Sources/Extension/UILabel+.swift
@@ -9,12 +9,11 @@ import UIKit
 
 extension UILabel {
     
-    func setPartialBold(boldText: String,fontSize: CGFloat){
-        let text = self.text ?? ""
-        let attributedString = NSMutableAttributedString(string: text)
-        let font = UIFont.systemFont(ofSize: 23, weight: .bold)
+    func setPartialBold(originalText: String,boldText: String,fontSize: CGFloat,boldFontSize: CGFloat){
+        let attributedString = NSMutableAttributedString(string: originalText)
+        let font = UIFont.systemFont(ofSize: boldFontSize, weight: .bold)
         self.font = UIFont.systemFont(ofSize: fontSize, weight: .light)
-        attributedString.addAttribute(.font,value: font,range: (text as NSString).range(of: boldText))
+        attributedString.addAttribute(.font,value: font,range: (originalText as NSString).range(of: boldText))
         self.attributedText = attributedString
     }
