+++ polaris-ios/polaris-ios/Sources/Util/StarNames.swift
@@ -1,0 +1,16 @@
+//
+//  StarNames.swift
+//  polaris-ios
+//
+//  Created by Yunjae Kim on 2021/04/20.
+//
+
+import Foundation
+
+struct StarNames {
+    static let starDict: Dictionary<String,String> = ["행복":"Happiness","절제":"Control",
+                                                      "감사":"Thanks","휴식":"Rest",
+                                                      "건강":"Health","성장":"Growth",
+                                                      "변화":"Change","극복":"Overcome",
+                                                      "도전":"Challenge"]
+}
